Added fix to Method Assignment

+++ MethodsAssignment/MethodsAssignment/Class1.cs
@@ -11,9 +11,9 @@ namespace MethodsAssignment
     internal class Class1
     {
         
-        public double math(int num1, [Optional] int num2)
+        public int math(int num1, [Optional] int num2)
         {
-            double totalAmount = num1 + num2 + 25;
+            int totalAmount = num1 + num2 + 25;
             return totalAmount;
         }
